Resolved merged conflicts
