+++ aether/client/components/header.jsx
@@ -1,19 +1,13 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 
 function Header() {
-
     return (
-     
-       
         <div id="header">
             <h1>Memory Report</h1>
             <div id="logotxt">
                 <img id="logo" src="/assets/AetherLogo01.png" alt="Aether Logo"/>
             </div>
-        </div>
-
-        
+        </div>   
     )
 }
-
 export default Header;
