+++ aether/index.html
@@ -6,7 +6,6 @@
 </head>
 <body>
   <div id="root"></div>
-
   <script src="build/bundle.js"></script>
 </body>
 </html>
