+++ aether/stylesheet.css
@@ -11,6 +11,13 @@ label {
     padding-right: 1rem;
 }
 
+select {
+    color: #5F6063;
+    background-color: white;
+    padding: 5px;
+    box-shadow: 1px 1px 1px 0px #d4d4d4;
+}
+
 nav a {
     padding: 1rem;
 }
@@ -40,12 +47,6 @@ nav a {
     padding: 0 1rem;
 }
 
-select {
-    color: #5F6063;
-    background-color: white;
-    padding: 5px;
-    box-shadow: 1px 1px 1px 0px #d4d4d4;
-}
 
 .options > div:nth-child(n + 2) {
     margin-left: 4rem;
