+++ aether/webpack.config.js
@@ -1,6 +1,5 @@
 const path = require('path');
 
-
 module.exports = {
   entry: path.resolve(__dirname, 'client/index.js'),
   output: {
@@ -11,7 +10,6 @@ module.exports = {
   devServer: {
     port: 5000,
     publicPath: '/build/',
-    // If proxy is needed
     proxy: {
       '/': 'http://localhost:9000',
       '/assets/': 'http://localhost:9000',
@@ -27,6 +25,4 @@ module.exports = {
       },
     ],
   },
-
-
 };
