+++ testing-sample-app/package.json
@@ -18,7 +18,6 @@
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "heapdump": "^0.3.15",
-    "heapsnapshot-parser": "^0.1.0",
     "nodemon": "^2.0.2",
     "parser": "^0.1.4",
     "path": "^0.12.7",
